ADD: auth for each lambda

+++ src/endpoints/products/deleteProduct.ts
@@ -1,4 +1,5 @@
 import { APIGatewayProxyEvent } from 'aws-lambda';
+import { isSeller } from 'src/lib/auth';
 import { DeleteProductResponse } from 'src/models/product-responses';
 import ProductRepository from 'src/repositories/ProductRepository';
 import Response from 'src/responses/Response';
@@ -9,18 +10,24 @@ async function deleteProduct(
   repository: ProductRepository,
   event: APIGatewayProxyEvent,
 ): Promise<Response> {
-  let response;
+  let response: Response;
 
-  try {
-    const { id } = event.pathParameters;
+  if (isSeller(event)) {
+    try {
+      const { id } = event.pathParameters;
 
-    repository.delete(id);
+      repository.delete(id);
 
-    response = new ResponseOk<DeleteProductResponse>();
-  } catch (error) {
+      response = new ResponseOk<DeleteProductResponse>();
+    } catch (error) {
+      response = new ResponseError({
+        message: 'Cannot find the product with that specific ID',
+      }, 404);
+    }
+  } else {
     response = new ResponseError({
-      message: 'Cannot find the product with that specific ID',
-    }, 404);
+      message: 'User not authorized',
+    }, 401);
   }
 
   return response;
